Polished Constants, and removed some naked variables
Constant "STOP_MOTOR" changed to "MOTOR_ZERO_SPEED" due to confusion with the Constant.

Clean up naked variables

Made Constants for Button IDs for 8 Buttons on an Xbox Controller. (More could be added if needed)

In a few subsystems, .setSmartCurrentLimit() often had a naked variable passed to it. Set all of them to Constants.MOTOR_CURRENT LIMIT, which equals 40. 40 was also the previous naked variable passed in theses subsystems